PT-1987-projects: Refactored entity group manager tree: 1. Renamed two types of groups to internal/external 2. Incorporated 'with parameter' property into the internal group manager (the WithType class removed) 3. Pulled AbstractPrimaryEntityGroup common to two group managers, handling operations unrelated to implementation.
