+++ components/entities/api/src/main/java/org/phenotips/entities/internal/AbstractInternalPrimaryEntityGroupManager.java
@@ -31,8 +31,10 @@ import org.xwiki.stability.Unstable;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
+import java.util.Map;
 
 import com.xpn.xwiki.doc.XWikiDocument;
 import com.xpn.xwiki.objects.BaseObject;
@@ -66,24 +68,23 @@ import com.xpn.xwiki.objects.BaseObject;
  * @since 1.3M2
  */
 @Unstable("New class and interface added in 1.3")
-public abstract class AbstractContainerPrimaryEntityGroup<G extends PrimaryEntity, E extends PrimaryEntity>
-    extends AbstractPrimaryEntityGroup<G, E>
+public abstract class AbstractInternalPrimaryEntityGroupManager<G extends PrimaryEntity, E extends PrimaryEntity>
+    extends AbstractExternalPrimaryEntityGroupManager<G, E>
     implements PrimaryEntityGroupManager<G, E>
 {
     /** The XClass used for storing membership information by default. */
     public static final EntityReference GROUP_MEMBER_CLASS =
-        new EntityReference("GroupMemberClass", EntityType.DOCUMENT,
-            Constants.CODE_SPACE_REFERENCE);
+        new EntityReference("GroupMemberClass", EntityType.DOCUMENT, Constants.CODE_SPACE_REFERENCE);
 
-    /** The XProperty used to save the class of contained members. */
-    protected static final String CLASS_XPROPERTY = "class";
-
-    protected AbstractContainerPrimaryEntityGroup(
-            EntityReference groupEntityReference, EntityReference memberEntityReference)
+    protected AbstractInternalPrimaryEntityGroupManager(
+        EntityReference groupEntityReference, EntityReference memberEntityReference)
     {
         super(groupEntityReference, memberEntityReference);
     }
 
+    /**
+     * Unlike getMembersMap, this will not read the parameters that may be saved with the members.
+     */
     @Override
     public Collection<E> getMembersOfType(G group, EntityReference type)
     {
@@ -176,7 +177,7 @@ public abstract class AbstractContainerPrimaryEntityGroup<G extends PrimaryEntit
         try {
             // FIXME GROUP_MEMBER_CLASS should be replaced with a static method call
             String bindingClassName = getLocalSerializer().serialize(
-                    AbstractContainerPrimaryEntityGroup.GROUP_MEMBER_CLASS);
+                    AbstractInternalPrimaryEntityGroupManager.GROUP_MEMBER_CLASS);
             String groupClass = getLocalSerializer().serialize(groupEntityReference);
             Query q = getQueryManager().createQuery(
                 ", BaseObject obj, BaseObject groupObj, StringProperty property "
@@ -201,5 +202,102 @@ public abstract class AbstractContainerPrimaryEntityGroup<G extends PrimaryEntit
                 ex.getMessage());
         }
         return Collections.emptyList();
+    }
+
+    /**
+     * Reads all members and their parameters. Returns a map with key: member name, value: parameters map (name:value).
+     * This function is not exposed to users of the group. If you need to save a member with parameters, override
+     * {@link setMemberParameters} and wrap this method. See {@link CollaboratorInProjectManager}.
+     *
+     * Saving members with parameters is used in cases where rebuilding them requires more information than is saved
+     * in the XWiki document. For example, a Collaborator is a User (which is a PrimaryEntity) combined with
+     * an AccessLevel. A Collaborator reference is saved together with information about creating the AccessLevel -
+     * this is handled in {@link CollaboratorInProjectManager.setMemberParameters()}. This method returns the
+     * parameters needed to rebuild a Collaborator, and the rebuilding of these objects is done in
+     * {@link CollaboratorInProjectManager.getMembers()}.
+     *
+     * @return parameters map
+     */
+    protected Map<String, Map<String, String>> getMembersMap(G group, EntityReference type)
+    {
+        try {
+            StringBuilder hql = new StringBuilder();
+
+            hql.append("select binding.value, propertyTable.name, propertyTable.value ")
+                .append("from BaseObject groupReference, StringProperty binding, BaseObject entity, ")
+                .append("     StringProperty propertyTable  ")
+                .append("where groupReference.id.id = binding.id and ")
+                .append("      binding.id.name = :referenceProperty and  ")
+                .append("      groupReference.name = :selfReference and  ")
+                .append("      entity.name = groupReference.name and ")
+                .append("      entity.id.id = propertyTable.id and ")
+                .append("      groupReference.className = :memberClass and ")
+                .append("      groupReference.number = entity.number and ")
+                .append("      binding.value in ");
+
+            // Subselect: names of all members of this document (of a certain type)
+            hql.append("(select distinct binding.value ")
+                .append(" from BaseObject groupReference, StringProperty binding, BaseObject entity, ")
+                .append("      StringProperty entityBinding ")
+                .append(" where groupReference.id.id = binding.id and ")
+                .append("       entity.name = groupReference.name and")
+                .append("       entity.id.id = entityBinding.id and")
+                .append("       groupReference.number = entity.number and")
+                .append("       binding.id.name = :referenceProperty and ")
+                .append("       groupReference.name = :selfReference and ")
+                .append("       entityBinding.id.name= :classProperty and ")
+                .append("       entityBinding.value = :entityType and ")
+                .append("       groupReference.className = :memberClass)");
+
+            Query q = getQueryManager().createQuery(hql.toString(), Query.HQL);
+
+            // FIXME
+            q.bindValue("selfReference", getFullSerializer().serialize(group.getDocument()).split(":")[1]);
+            q.bindValue("memberClass", getLocalSerializer().serialize(GROUP_MEMBER_CLASS));
+            q.bindValue("entityType", getLocalSerializer().serialize(type));
+            q.bindValue("referenceProperty", getMembershipProperty());
+            q.bindValue("classProperty", CLASS_XPROPERTY);
+
+            List<Object> members = q.execute();
+
+            return this.buildParametersMap(members);
+        } catch (QueryException ex) {
+            this.logger.warn("Failed to query members: {}", ex.getMessage());
+        }
+        return Collections.emptyMap();
+    }
+
+    /**
+     * If an extending class wants to record parameters with a member, override this.
+     *
+     * @param member that is saved
+     * @param obj object for recording parameters
+     */
+    protected void setMemberParameters(E member, BaseObject obj)
+    {
+        // do nothing as default.
+    }
+
+    private Map<String, Map<String, String>> buildParametersMap(List<Object> members)
+    {
+        Map<String, Map<String, String>> parameters = new HashMap<>();
+
+        for (Object o : members) {
+            Object[] asArray = (Object[]) o;
+
+            String entityName = (String) asArray[0];
+            String propertyName = (String) asArray[1];
+            String propertyValue = (String) asArray[2];
+
+            Map<String, String> map = parameters.get(entityName);
+            if (map == null) {
+                map = new HashMap<>();
+                parameters.put(entityName, map);
+            }
+
+            map.put(propertyName, propertyValue);
+        }
+
+        return parameters;
     }
 }
